Extract unavailable-tool notification into its own module
Move the VS Code notification logic (showInformationMessage + session
dedup) out of runToolUseFlow.ts into src/tools/toolUnavailableNotification.ts.
This decouples the flow runner from the vscode UI layer — the vscode
import is now gone from runToolUseFlow.ts entirely.
